Removed no longer used module GeneAnnotation.
This module has been causing lots of problems with local installs.
The core files to this module module do not even exist.
Please, if any developer no longer needs a module, remove it from the baseline.
DO NOT simply remove the underlying files in the module directory.
This is much much worse as it screws up the drupal database and causes all kinds of errors!